Fix RpcOutputRequestProcessor to correctly reply ok message
RpcOutputRequestProcessor was incorrectly interpreting
if reponseOutput is empty or not. Check for emptyness
of rpc output itself was added, if empty ok element is
created for output. This is according to YANG RFC 7950,
if rpc output is empty, ok element is returned.

Signed-off-by: Martin Bugan <martin.bugan@pantheon.tech>
(cherry picked from commit c5e5802e14e9c998caffd9355a456475ab14e032)